requirements.txt
aiohttp==3.9.3
python-dotenv==1.0.1
tqdm==4.66.2
